implement data and bss ops and memory loading
required some interface changes for Env as well.

these are incredibly useful if not necessary operations to have for the
vm opcode set. I intended to also use these for the C frontend, but it
seems that my ideas for the vm have diverged enough from chibicc's
original design that this won't be easy.

this might be a natural stopping point for the frontend piece of this
project. I think I'm going to call it here on the C side and move on
to making an IR for the instruction set.